V0.0.0.19 Pre-Alpha (Day)
	Implementations
	- Updated Alarm 0 in core game to allow the transition from Sequence to Battle UI to happen in endgame
	- Created Objects for moves that will appear in the sequence for attacks.
	- Implemented Fix for Zero issue by reversing the array checker
	- Updated d20Visual into diceVisual that handles all dice visual rolls
	- Making D4, D6, D8, D10, D12 objects. + with Updates on Sprites to have Numbers
	- Moved all D20 Object code into obj_diceRollParent for easy reuse with the other dice by making a new Variable Definition "topEnd" to hold the original dice top end.
	- Imported Mega Lucario Sprite and Added it to Title Screen
	- Updated HP Code to have health number moved over to the left to get full visual in endGame
	- Began developing new version of move system. However, might skip until after PlayTest. (If it can't work.)

Bugs/Issues
The reason why it adds zeros (Even through I don't get it), Every zero is a skipped move which is why the zero was put before the Work up. Now it does it for all of the endGame Moves. I have no idea why it does this and it is really frustrating. (Might have to consider bringing this inside the engine if it keeps this up after the prototype)